Replaced SystemDynamicTemplateModel with alternative
Depending on the system is considered to be bad practice, so by
splitting the example page from its menu-items into separate components,
we break the dependency cycle and avoid the need to use the system.

The systemref is still established and passed to the system constructor,
because this may be useful for development tools.