Replace WebService
WebService has been removed. It has been superceded by the RouteProvider
protocol in bidi 1.18.x.

Bidi's IdentifiableHandler provides for a better way of using keywords
with path-for.

The original bidi router in this modular became too complex. One thing
that has been lost in the simplification is the component rethrow of
exceptions. It is possible that a future commit will restore this
feature.